Make derivations compatible with aeson
This is a large patch that is intended to make derivations as close as
those from Data.Aeson as possible.

It replaces the custom settings from `elm-bridge` with that of `aeson`.

The derived elm code currently requires a few helper functions that
reside in the `examples/Helpers.elm` file. This was decided to make the
template Haskell code simpler and easier to generate and to have quick
results.

This patch also introduces a big `QuickCheck` based test to increase
confidence in the quality of the code. It generates an Elm file that
runs tests.

It also temporarily locks the package to aeson 0.9 because of this bug:
https://github.com/bos/aeson/issues/293